MLP model implemenation with @leopoldoagorio (#29)
